Moved some tests. Trying to route normal api
