+++ Fixtures/IntegrationTests/IntegrationTests/BandwidthMetricEvents.swift
@@ -1,0 +1,14 @@
+//
+//  BandwidthMetricEvents.swift
+//  IntegrationTests
+//
+//  Created by Santiago Puppo on 4/7/25.
+//
+
+import Testing
+
+@Suite
+struct BandwidthMetricEvents {
+    
+    
+}
